requirements.txt
fastapi
uvicorn[standard]
python-multipart  
Pillow
torch
docling_core
git+https://github.com/huggingface/transformers
streamlit 
requests         
